v1 of both fixes

+++ testnet/run.sh
@@ -1,5 +1,4 @@
 #!/usr/bin/env bash
-set -e
 
 # https://github.com/stakater/til/blob/master/ibm/ibm-pvc-lost-found-directory-issue.md
 # On some systems PVCs will be mounted with a lost+found folder so it is not neccessarily
@@ -30,19 +29,20 @@ fi
 
 echo "Executing command tendermint $@"
 
-tendermint $@ || true
-
-echo "\n\n\n THIS FAILED! Attempting file corruption healing and a restart"
-
-#sleep 999
-
-echo "running wal2json"
-./wal2json ./data/cs.wal/wal > wal.json || true
-
-echo "running json2wal to replace data wal file"
-rm -rf ./data/cs.wal/wal
-./json2wal wal.json ./data/cs.wal/wal || true
-
-echo "restarting tendermint"
 tendermint $@
 
+if [ $? == 33 ]; then
+    echo -e "\n\nTHIS FAILED! Attempting file corruption healing and a restart"
+
+    echo "running wal2json"
+    ./wal2json ./data/cs.wal/wal > wal.json
+
+    echo "running json2wal to replace data wal file"
+    rm -rf ./data/cs.wal/wal
+    ./json2wal wal.json ./data/cs.wal/wal
+
+    echo "restarting tendermint"
+    tendermint $@
+else
+    echo "Failed with code $?"
+fi
